+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+
+Features:
+- **BREAKING**: `async` subscribers are now supported. The `EventDispatcher.dispatch` returns a future. 
+
 ## 1.0.0
 
 - Initial version.